refactor to make code easier to read and more modular add TODOs for work on custom response schema
