+++ gpt/response.go
@@ -1,0 +1,22 @@
+package gpt
+
+type GptResponse struct {
+	ID      string    `json:"id"`
+	Object  string    `json:"object"`
+	Created int64     `json:"created"`
+	Model   string    `json:"model"`
+	Choices []Choice  `json:"choices"`
+	Usage   UsageInfo `json:"usage"`
+}
+
+type Choice struct {
+	Index        int        `json:"index"`
+	Message      GptMessage `json:"message"`
+	FinishReason string     `json:"finish_reason"`
+}
+
+type UsageInfo struct {
+	PromptTokens     int `json:"prompt_tokens"`
+	CompletionTokens int `json:"completion_tokens"`
+	TotalTokens      int `json:"total_tokens"`
+}
