+++ tests/integration_test.go
@@ -1,4 +1,4 @@
-package tests
+package integration_test
 
 import (
 	"strings"
